test(accordion): restore 5s lazy-load wait to fix intersect-load flake
TestAccordion_ServerLoadedContent and TestIntegration/Full_Workflow flaked
in CI ('Timeout 2000ms exceeded / server-loaded content should appear').
The accordion lazy body loads via hx-trigger='intersect once' (IntersectionObserver
+ GET + innerHTML swap); commit d3d5cf3 tightened the WaitFor from 5000ms to
2000ms (comment still said 'max 5 seconds'), too tight for the intersect+network
chain under full-suite browser contention. Restore all three lazy-load waits to
5000ms (proven stable pre-regression).

+++ tests/e2e/components_test.go
@@ -114,7 +114,7 @@ func TestAccordion_ServerLoadedContent(t *testing.T) {
 		// Wait for HTMX to fetch content (max 5 seconds)
 		err = page.Locator("text=Server Response A").WaitFor(playwright.LocatorWaitForOptions{
 			State:   playwright.WaitForSelectorStateVisible,
-			Timeout: playwright.Float(2000),
+			Timeout: playwright.Float(5000),
 		})
 		require.NoError(t, err, "server-loaded content should appear")
 
@@ -156,7 +156,7 @@ func TestAccordion_ServerLoadedContent(t *testing.T) {
 		// Wait for content to load
 		err = page.Locator("text=Server Response B").WaitFor(playwright.LocatorWaitForOptions{
 			State:   playwright.WaitForSelectorStateVisible,
-			Timeout: playwright.Float(2000),
+			Timeout: playwright.Float(5000),
 		})
 		require.NoError(t, err)
 
@@ -653,7 +653,7 @@ func TestIntegration(t *testing.T) {
 		// Wait for content
 		err = page.Locator("text=Server Response A").WaitFor(playwright.LocatorWaitForOptions{
 			State:   playwright.WaitForSelectorStateVisible,
-			Timeout: playwright.Float(2000),
+			Timeout: playwright.Float(5000),
 		})
 		require.NoError(t, err)
